Move Noise poll to new file, and use common poll
